completed the implementation of the usage next going to separate the main service into usage-service auth-service main-service

+++ main-service/src/main/java/com/quotaGate/main_service/Service/TokenService.java
@@ -70,7 +70,7 @@ public class TokenService {
 
         JwtDTO jwtDTO = new JwtDTO();
         jwtDTO.setEmail(claims.get("email", String.class));
-        jwtDTO.setId(claims.get("id", Integer.class));
+        jwtDTO.setId(claims.get("id", Long.class));
 
         return jwtDTO;
     }
